+++ packages/tswebextension/test/lib/mv2/background/tabs/frame.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import { type MatchingResult, type CosmeticResult } from '@adguard/tsurlfilter';
 
+import { MAIN_FRAME_ID, NO_PARENT_FRAME_ID } from '../../../../../src/lib/common/constants';
 import { FrameMV2 } from '../../../../../src/lib/mv2/background/tabs/frame';
 
 describe('Frame', () => {
@@ -10,7 +11,8 @@ describe('Frame', () => {
             const cosmeticResult = {} as CosmeticResult;
             const matchingResult = {} as MatchingResult;
             const tabId = 1;
-            const frameId = 0;
+            const frameId = MAIN_FRAME_ID;
+            const parentFrameId = NO_PARENT_FRAME_ID;
             const timeStamp = Date.now();
             const parentDocumentId = '1';
 
@@ -18,6 +20,7 @@ describe('Frame', () => {
                 url,
                 tabId,
                 frameId,
+                parentFrameId,
                 timeStamp,
                 parentDocumentId,
             });
@@ -27,6 +30,7 @@ describe('Frame', () => {
             expect(frame).toBeInstanceOf(FrameMV2);
             expect(frame.url).toBe(url);
             expect(frame.frameId).toBe(frameId);
+            expect(frame.parentFrameId).toBe(parentFrameId);
             expect(frame.tabId).toBe(tabId);
             expect(frame.cosmeticResult).toBe(cosmeticResult);
             expect(frame.matchingResult).toBe(matchingResult);